Move Test functions to separate files.
