+++ cdh-tsat6-stm32project-boardV1/Core/Src/main.c
@@ -26,7 +26,6 @@
 #include <stdlib.h>
 
 #include "W25N_driver.h"
-#include "camera_driver.h"
 #include "W25N_driver_test.h"
 #include "AS3001204_driver.h"
 #include "AS3001204_driver_test.h"
@@ -34,6 +33,9 @@
 #include "MAX6822_driver.h"
 #include "LTC1154_driver.h"
 #include "can.h"
+
+#include "camera_driver.h"
+#include "camera_driver_test.h"
 /* USER CODE END Includes */
 
 /* Private typedef -----------------------------------------------------------*/
