+++ cdh-tsat6-stm32project-nucleo/Core/Src/main.c
@@ -30,13 +30,15 @@
 #include "W25N_driver.h"
 #include "W25N_driver_test.h"
 
-#include "camera_driver.h"
 #include "AS3001204_driver.h"
 #include "AS3001204_driver_test.h"
 #include "LEDs_driver.h"
 #include "MAX6822_driver.h"
 #include "LTC1154_driver.h"
 #include "can.h"
+
+#include "camera_driver.h"
+#include "camera_driver_test.h"
 /* USER CODE END Includes */
 
 /* Private typedef -----------------------------------------------------------*/
