+++ cdh-tsat6-stm32project-nucleo/Drivers/Hardware_Peripherals/Src/camera_driver.c
@@ -286,18 +286,3 @@ void enable_piCAM_UART()
     HAL_NVIC_EnableIRQ(piCAM_UART_IRQn);
 }
 
-/*----------------------------------------------------------------------------------------------
-Public Testing Function Definitions
------------------------------------------------------------------------------------------------*/
-void piCAM_Test_Procedure()
-{
-    piCAM_Boot_Up_Sequence();
-
-    HAL_Delay(3000);
-    piCAM_Status_Test();
-
-    HAL_Delay(3000);
-    piCAM_Capture_Daylight();
-
-    piCAM_DMA_Start();
-}
